Add module to get symbol listing Move unit tests for symbols check to a new test script. Add unit test condition to check for Company name from symbol listing
